clean up and fix NRE
