+++ src/OrchardCore/OrchardCore.Users.Core/Models/RegistrationOptions.cs
@@ -2,8 +2,6 @@ namespace OrchardCore.Users.Models;
 
 public class RegistrationOptions
 {
-    public bool AllowSiteRegistration { get; set; }
-
     public bool UsersMustValidateEmail { get; set; }
 
     public bool UsersAreModerated { get; set; }
